requirements.txt
langchain>=0.0.350
neo4j>=5.14.1
python-dotenv>=1.0.0
openai>=1.3.7
tiktoken>=0.5.1
jupyter>=1.0.0
notebook>=7.0.0
